UiElementWithBackground made abstract
