+++ src/viewModel/GeneralAbstracts/GuiElem.ts
@@ -1,3 +1,5 @@
+import * as PIXI from 'pixi.js';
+
 export abstract class GuiElem {
   abstract pixiElem: PIXI.Container;
 }
